# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Order POS API",
        "description": "API para la gestión de pedidos, productos y usuarios en puntos de venta.",
        "contact": {
            "name": "Jhonatan Guerrero",
            "email": "jhonatanguerrero@outlook.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener todos los usuarios activos",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios activos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Crear un nuevo usuario",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jhonatan Guerrero"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "jguerrero"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
...
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener usuario por ID",
                "operationId": "36a33ff774d5cba33c039dec2c3e0287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar usuario",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nuevo Nombre"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "nuevo_usuario"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "nuevaPassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "nuevaPassword123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar usuario (soft delete)",
                "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        },
        "/api/users/name/{name}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Buscar usuarios por nombre",
                "operationId": "595d3f734556daf38a301a4477ad7583",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Nombre o parte del nombre",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuarios encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron usuarios"
                    }
                }
            }
        },
        "/api/users/{id}/restore": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Restaurar usuario eliminado",
                "operationId": "a3087db765f18596ec09f9c3abb4fdc5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario restaurado correctamente"
                    },
                    "404": {
                        "description": "Usuario eliminado no encontrado"
                    }
                }
            }
        },
        "/api/users/{id}/force": {
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar usuario permanentemente",
                "operationId": "5eeae45fccc78c2a9f675609634f785a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado permanentemente"
                    },
                    "404": {
                        "description": "Usuario eliminado no encontrado"
                    }
                }
            }
        },
        "/api/users/trashed/{id}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener usuario eliminado por ID",
                "operationId": "77311b1003788e5469b922a37aa6db17",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario eliminado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario eliminado no encontrado"
                    }
                }
            }
        },
        "/api/users/{id}/permissions": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Asignar o revocar permisos a un usuario",
                "operationId": "5e39cd26d8be973cfb25d30ec4007892",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "permissions",
                                    "assign"
                                ],
                                "properties": {
                                    "permissions": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "edit-users"
                                        }
                                    },
                                    "assign": {
                                        "description": "true para asignar, false para revocar",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permisos actualizados"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "Usuario",
                "description": "Modelo que representa un usuario del sistema",
                "required": [
                    "id",
                    "name",
                    "username"
                ],
                "properties": {
                    "id": {
                        "description": "ID único del usuario",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Nombre completo del usuario",
                        "type": "string",
                        "example": "Jhonatan Guerrero"
                    },
                    "username": {
                        "description": "Nombre de usuario para login",
                        "type": "string",
                        "example": "jguerrero"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-03T12:34:56Z"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-03T12:34:56Z"
                    },
                    "deleted_at": {
                        "description": "Fecha de eliminación (null si activo)",
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    },
                    "roles": {
                        "description": "Roles asignados al usuario",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "admin"
                        }
                    },
                    "permissions": {
                        "description": "Permisos asignados al usuario",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "edit-users"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}